+++ R/sample_mallows.R
@@ -42,7 +42,7 @@
 #' @example /inst/examples/sample_mallows_example.R
 #'
 sample_mallows <- function(rho0, alpha0, n_samples,
-                           leap_size = 1,
+                           leap_size = max(1L, floor(n_items/5)),
                            metric = "footrule",
                            diagnostic = FALSE,
                            burnin = ifelse(diagnostic, 0, 1000),
@@ -80,7 +80,8 @@ sample_mallows <- function(rho0, alpha0, n_samples,
     burnin = internal_burnin,
     thinning = internal_thinning,
     leap_size = leap_size,
-    metric = metric
+    metric = metric,
+    obs_freq = rep(1, internal_n_samples)
   ))
 
   if(diagnostic){