+++ src/distances.cpp
@@ -50,7 +50,7 @@ double spearman_distance(const arma::vec& r1, const arma::vec& r2){
   return std::pow(arma::norm(r1 - r2, 2), 2.0);
 }
 
-double ulam_distance (const arma::vec& r1, const arma::vec& r2){
+double ulam_distance(const arma::vec& r1, const arma::vec& r2){
 
   int N = r1.n_elem;
 
@@ -75,13 +75,13 @@ double ulam_distance (const arma::vec& r1, const arma::vec& r2){
 }
 
 
-//' Compute the distance between two rank vectors.
+//' Compute the Distance between two rankings
 //'
+//' @description Compute the distance between two rankings according to several metrics.
 //' @param r1 A vector of ranks.
 //' @param r2 A vector of ranks.
 //' @param metric A string. Available options are \code{"footrule"},
 //' \code{"kendall"}, \code{"cayley"}, \code{"hamming"}, \code{"spearman"}, and \code{"ulam"}.
-//' Defaults to \code{"footrule"}.
 //' @return A scalar.
 //' @details Note that the Spearman distance is the squared L2 norm, whereas
 //' the footrule distance is the L1 norm.
@@ -92,12 +92,10 @@ double ulam_distance (const arma::vec& r1, const arma::vec& r2){
 //'
 //'
 //' @references \insertAllCited{}
-//'
 //' @keywords internal
 //'
-//'
 // [[Rcpp::export]]
-double get_rank_distance(arma::vec r1, arma::vec r2, std::string metric){
+double  get_rank_distance(arma::vec r1, arma::vec r2, std::string metric){
 
   if (r1.n_elem != r2.n_elem){
     Rcpp::stop("r1 and r2 must have the same length");
@@ -121,22 +119,29 @@ double get_rank_distance(arma::vec r1, arma::vec r2, std::string metric){
 }
 
 
-// Compute the distance between all rows in rankings and rho, and return the sum
-double rank_dist_sum(const arma::mat& rankings, const arma::vec& rho, const std::string& metric){
-  return arma::sum(rank_dist_vec(rankings, rho, metric));
+
+// [[Rcpp::export]]
+double rank_dist_sum(const arma::mat& rankings, const arma::vec& rho,
+                     const std::string& metric, const arma::vec& obs_freq){
+  return arma::sum(rank_dist_vec(rankings, rho, metric, obs_freq));
 }
 
 
-// Compute the distance between each assessor's ranking and the consensus
-arma::vec rank_dist_vec(const arma::mat& rankings, const arma::vec& rho,
-                                 const std::string& metric){
+
+// [[Rcpp::export]]
+arma::vec rank_dist_vec(const arma::mat& rankings,
+                        const arma::vec& rho,
+                        const std::string& metric,
+                        const arma::vec& obs_freq){
 
   int n = rankings.n_cols;
   arma::vec result = arma::zeros(n);
 
   for(int i = 0; i < n; ++i){
-    result(i) = get_rank_distance(rankings.col(i), rho, metric);
+    result(i) = get_rank_distance(rankings.col(i), rho, metric) * obs_freq(i);
   }
   return(result);
 }
+
+