+++ NEWS.md
@@ -1,3 +1,11 @@
+# Bayes Mallows 1.0.0
+
+* Function for computing likelihood added.
+* Options for dealing with missing values added, and documentation now states how missing values are dealt with.
+* Function rank_freq_distr added, for computing the frequency distribution of ranking patterns.
+* Function rank_distance added, for computing the distance between rankings.
+* Function expected_dist added for computing expectation of several metrics under the Mallows model.
+
 # BayesMallows 0.5.0
 
 * Function compute_consensus now includes an option for computing consensus of augmented ranks.